streamlit_app/requirements.txt
streamlit
langchain
langgraph
requests
beautifulsoup4
PyYAML
pydantic
google-generativeai
openai
streamlit-agraph
termcolor
